+++ packages/sdk/CHANGELOG.md
@@ -2,7 +2,8 @@
 
 - YYYY-MM-DD **BREAKING?** -- description
 
-- 2021-04-20 -- Removes **Deprecated** `params`, `buildParams`
+- 2021-04-22 -- Adds `wallet-utils` `validateSignableTransaction` support for wallets to validate Signable payload
+- 2021-04-21 -- Removes **Deprecated** `params`, `buildParams`
 - 2021-04-21 -- Updates encoding naming of `gasLimit` and `script` to `computeLimit` and `cadence`. Internal only.
 
 ### 0.0.45-alpha.20 -- 2021-04-21